Extract tests with S3 resources from TestGemRemoteFetcher
